requirements.txt
annotated-types==0.6.0
anyio==4.2.0
assemblyai==0.20.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
python-telegram-bot==20.7
requests==2.31.0
sniffio==1.3.0
sounddevice==0.4.6
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
websockets==12.0
